requirements_dispatch.txt
psycopg2-binary>=2.9.0
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.20
python-dotenv>=1.0.0
APScheduler>=3.10.0
requests>=2.31.0
geopy>=2.3.0

